requirements.txt
streamlit
pillow
scikit-learn
matplotlib
numpy
kagglehub
